Move and update test with real c++ implementation
