--- src/tests/JIT/Directed/pinvoke/pinvoke-allPlatforms.csproj
@@ -1,10 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <RequiresProcessIsolation>true</RequiresProcessIsolation>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="pinvoke-allPlatforms.cs" />
-    <Compile Include="GameControllerButtonBind.cs" />
-  </ItemGroup>
-</Project>
